--- src/main/java/com/socialsitebackend/socialsite/controller/UserController.java
@@ -1,69 +1,0 @@
-package com.socialsitebackend.socialsite.controller;
-
-import com.socialsitebackend.socialsite.dto.AuthToken;
-import com.socialsitebackend.socialsite.dto.LoginRequestDto;
-import com.socialsitebackend.socialsite.dto.UserRegisterRequestDto;
-import com.socialsitebackend.socialsite.model.User;
-import com.socialsitebackend.socialsite.security.TokenProvider;
-import com.socialsitebackend.socialsite.service.UserService;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.web.bind.annotation.*;
-
-@CrossOrigin(origins = "*", maxAge = 3600)
-@RestController
-@RequestMapping("/users")
-public class UserController {
-
-    private final AuthenticationManager authenticationManager;
-    private final TokenProvider jwtTokenUtil;
-    private final UserService userService;
-
-    public UserController(AuthenticationManager authenticationManager, TokenProvider jwtTokenUtil, UserService userService) {
-        this.authenticationManager = authenticationManager;
-        this.jwtTokenUtil = jwtTokenUtil;
-        this.userService = userService;
-    }
-
-    @RequestMapping(value = "/authenticate", method = RequestMethod.POST)
-    public ResponseEntity<?> generateToken(@RequestBody LoginRequestDto loginUser) throws AuthenticationException {
-
-        final Authentication authentication = authenticationManager.authenticate(
-                new UsernamePasswordAuthenticationToken(
-                        loginUser.getUsername(),
-                        loginUser.getPassword()
-                )
-        );
-        SecurityContextHolder.getContext().setAuthentication(authentication);
-        final String token = jwtTokenUtil.generateToken(authentication);
-        return ResponseEntity.ok(new AuthToken(token));
-    }
-
-    @RequestMapping(value="/register", method = RequestMethod.POST)
-    public User saveUser(@RequestBody UserRegisterRequestDto user){
-        return userService.save(user);
-    }
-
-
-    @PreAuthorize("hasRole('ADMIN')")
-    @RequestMapping(value="/adminping", method = RequestMethod.GET)
-    public String adminPing(){
-        return "Only Admins Can Read This";
-    }
-
-    @PreAuthorize("hasRole('USER')")
-    @RequestMapping(value="/userping", method = RequestMethod.GET)
-    public String userPing(){
-        return "Any User Can Read This";
-    }
-
-    @RequestMapping(value="/usernew", method = RequestMethod.GET)
-    public String userTest(){
-        return "Any User Can Read This";
-    }
-}
