--- src/main/java/com/socialsitebackend/socialsite/dto/AddPostDto.java
@@ -1,12 +1,0 @@
-package com.socialsitebackend.socialsite.dto;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Builder
-@Data
-public class AddPostDto {
-    private String title;
-    private String description;
-    private String photoUrl;
-}
