--- src/main/java/com/socialsitebackend/socialsite/dto/AuthToken.java
@@ -1,10 +1,0 @@
-package com.socialsitebackend.socialsite.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class AuthToken {
-    private String token;
-}
