--- src/main/java/com/socialsitebackend/socialsite/dto/LoginRequestDto.java
@@ -1,11 +1,0 @@
-package com.socialsitebackend.socialsite.dto;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Builder
-@Data
-public class LoginRequestDto {
-    private String username;
-    private String password;
-}
