+++ src/main/java/com/socialsitebackend/socialsite/entities/Role.java
@@ -1,0 +1,20 @@
+package com.socialsitebackend.socialsite.entities;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Entity
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Role {
+    @Id
+    @GeneratedValue
+    private long id;
+    private String name;
+    private String description;
+
+}
