--- src/main/java/com/socialsitebackend/socialsite/model/Role.java
@@ -1,20 +1,0 @@
-package com.socialsitebackend.socialsite.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Role {
-    @Id
-    @GeneratedValue
-    private long id;
-    private String name;
-    private String description;
-
-}
