+++ src/main/java/com/socialsitebackend/socialsite/post/PostRepository.java
@@ -1,0 +1,12 @@
+package com.socialsitebackend.socialsite.post;
+
+import com.socialsitebackend.socialsite.entities.PostEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+
+
+@Repository
+public interface PostRepository extends JpaRepository<PostEntity,Long> {
+
+}
