--- src/main/java/com/socialsitebackend/socialsite/repository/PostRepository.java
@@ -1,12 +1,0 @@
-package com.socialsitebackend.socialsite.repository;
-
-import com.socialsitebackend.socialsite.model.Post;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-
-
-@Repository
-public interface PostRepository extends JpaRepository<Post,Long> {
-
-}
