--- src/main/java/com/socialsitebackend/socialsite/repository/RoleRepository.java
@@ -1,12 +1,0 @@
-package com.socialsitebackend.socialsite.repository;
-
-
-import com.socialsitebackend.socialsite.model.Role;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
-    Role findRoleByName(String name);
-
-}
