--- src/main/java/com/socialsitebackend/socialsite/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.socialsitebackend.socialsite.repository;
-
-
-
-import com.socialsitebackend.socialsite.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    User findByUsername(String username);
-
-}
