+++ src/main/java/com/socialsitebackend/socialsite/role/RoleRepository.java
@@ -1,0 +1,12 @@
+package com.socialsitebackend.socialsite.role;
+
+
+import com.socialsitebackend.socialsite.entities.Role;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RoleRepository extends JpaRepository<Role, Long> {
+    Role findRoleByName(String name);
+
+}
