--- src/main/java/com/socialsitebackend/socialsite/security/CORSFilter.java
@@ -1,22 +1,0 @@
-package com.socialsitebackend.socialsite.security;
-
-import javax.servlet.*;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-public class CORSFilter implements Filter {
-
-    public void doFilter(ServletRequest req, ServletResponse res, FilterChain chain) throws IOException, ServletException {
-        HttpServletResponse response = (HttpServletResponse) res;
-        response.setHeader("Access-Control-Allow-Origin", "*");
-        response.setHeader("Access-Control-Allow-Credentials", "true");
-        response.setHeader("Access-Control-Allow-Methods", "POST, GET, PUT, OPTIONS, DELETE");
-        response.setHeader("Access-Control-Max-Age", "3600");
-        response.setHeader("Access-Control-Allow-Headers", "X-Requested-With, Content-Type, Authorization, Origin, Accept, Access-Control-Request-Method, Access-Control-Request-Headers");
-        chain.doFilter(req, res);
-    }
-
-    public void init(FilterConfig filterConfig) {}
-
-    public void destroy() {}
-
-}
