--- src/main/java/com/socialsitebackend/socialsite/security/UnauthorizedEntryPoint.java
@@ -1,21 +1,0 @@
-package com.socialsitebackend.socialsite.security;
-
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.io.Serializable;
-
-
-@Component
-public class UnauthorizedEntryPoint implements AuthenticationEntryPoint, Serializable {
-
-    @Override
-    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException {
-        response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "Unauthorized");
-    }
-
-}
