+++ src/main/java/com/socialsitebackend/socialsite/user/UserRepository.java
@@ -1,0 +1,13 @@
+package com.socialsitebackend.socialsite.user;
+
+
+
+import com.socialsitebackend.socialsite.entities.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
+    UserEntity findByUsername(String username);
+
+}
