+++ src/main/java/com/socialsitebackend/socialsite/user/dto/AuthToken.java
@@ -1,0 +1,10 @@
+package com.socialsitebackend.socialsite.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class AuthToken {
+    private String token;
+}
