+++ src/main/java/com/socialsitebackend/socialsite/user/dto/UserLoginRequestDto.java
@@ -1,0 +1,11 @@
+package com.socialsitebackend.socialsite.user.dto;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Builder
+@Data
+public class UserLoginRequestDto {
+    private String username;
+    private String password;
+}
