+++ src/main/java/com/socialsitebackend/socialsite/user/dto/UserRegisterRequestDto.java
@@ -1,0 +1,16 @@
+package com.socialsitebackend.socialsite.user.dto;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Builder
+@Data
+public class UserRegisterRequestDto {
+    private String username;
+    private String password;
+    private String email;
+
+    private String avatarUrl;
+    private String gender;
+    private String profileDescription;
+}
